Extract EquiJoinPlanner
